requirements.txt
PyYAML
pandas
joblib
flask
watchdog